DF-10716 fix warnings as errors on win7 for unit tests
Former-commit-id: 5253d5afea4534b9c222b224c3d5fe9b71fe669e

+++ Projects/UnitTests/Classes/Tests/LoadImageTest.cpp
@@ -132,7 +132,7 @@ DAVA_TESTCLASS (LoadImageTest)
             TEST_VERIFY(loadedMip->height == expectedMipHeight);
             TEST_VERIFY(loadedMip->width == expectedMipWidth);
 
-            uint32 sampleSize = 0;
+            uint64 sampleSize = 0;
             uint8* sampleData = nullptr;
             FilePath sampleMipPath(Format("~res:/TestData/LoadImageTest/dxt1_mip%u.dat", i));
             sampleData = FileSystem::Instance()->ReadFileContents(sampleMipPath, sampleSize);
